requirements.txt
filelock==3.18.0
fvcore==0.1.5.post20221221
iopath==0.1.10
numpy==2.2.6
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
scikit_learn==1.6.1
simplejson==3.19.3
termcolor==3.1.0
timm==1.0.15
torch==2.5.1+cu124
torchvision==0.20.1+cu124
